+++ packages/go-kosu/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 
 -   Bump tendermint version from v0.32.3 to v0.32.4
+-   Sort ValidatorUpdates by address and returns updated set
 -   Fix witness unsupported key logger bug
 -   Upgrade tendermint/tm-db to v0.2.0
 -   Fix witness connection mngmt and error handling